+++ src/Serializers/RBX.php
@@ -179,6 +179,6 @@ class RBX implements Serializer
      */
     public function __toString() : string
     {
-        return 'RBX';
+        return "RBX (base: {$this->base})";
     }
 }